refactor: reorder hooks
